requirements.txt
aiogram==2.21
python-dotenv==0.20.0
requests==2.28.0